backend/requirements/requirements.txt
adb==1.2.0
archinfo==7.7.9.14
bitstring==3.1.5
capstone-windows==3.0.4
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
Django==2.0rc1
django-jinja==2.4.0
djangorestframework==3.7.3
idna==2.6
Jinja2==2.9.6
libusb1==1.6.4
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
pycparser==2.18
pygame==1.9.3
pytz==2017.3
requests==2.18.4
six==1.11.0
style==1.1.0
typing==3.6.2
update==0.0.1
urllib3==1.22
